server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "graxel"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

